+++ mobile2/src/main/java/sopt/mobile2/dto/AccountResponseDto.java
@@ -2,7 +2,7 @@ package sopt.mobile2.dto;
 
 import sopt.mobile2.domain.Account;
 
-public record AccountResponseDto(String accountName, int balance, Long accountId) {
+public record AccountResponseDto(String accountName, long balance, Long accountId) {
 
     public static AccountResponseDto of(Account account) {
         return new AccountResponseDto(account.getAccountName(), account.getBalance(), account.getId());
