+++ mobile2/src/main/java/sopt/mobile2/dto/MonthlyTransferResponseDto.java
@@ -1,5 +1,5 @@
 package sopt.mobile2.dto;
 
 public record MonthlyTransferResponseDto(String accountName, String date,
-                                         int transferAmount, int balance, boolean isWithdraw, String hashTag) {
+                                         long transferAmount, long balance, boolean isWithdraw, String hashTag) {
 }
