+++ mobile2/src/main/java/sopt/mobile2/dto/PaymentResponseDto.java
@@ -2,9 +2,9 @@ package sopt.mobile2.dto;
 
 import java.util.List;
 
-public record PaymentResponseDto(int payment, boolean isTotalOverZero, List<MonthlyTransferResponseDto> monthlyTransferList) {
+public record PaymentResponseDto(long payment, boolean isTotalOverZero, List<MonthlyTransferResponseDto> monthlyTransferList) {
 
-    public static PaymentResponseDto of(int payment, boolean isTotalOverZero, List<MonthlyTransferResponseDto> monthlyTransferList) {
+    public static PaymentResponseDto of(long payment, boolean isTotalOverZero, List<MonthlyTransferResponseDto> monthlyTransferList) {
         return new PaymentResponseDto(payment, isTotalOverZero, monthlyTransferList);
     }
 }
